publish company variable names fixed
