+++ controllers/publishCompanyController.js
